Made a common ScheduleSearchForm and ScheduleView class for various schedule routes.
Adjusted ScheduleTemplate to allow showing referee details for the assignor.
