+++ src/AppBundle/Action/Schedule2016/routes.yml
@@ -18,6 +18,16 @@ schedule_official_2016:
         _view_csv:   schedule_official_view_csv_2016
         _view_txt:   schedule_official_view_txt_2016
 
+schedule_assignor_2016:
+    path: /assignor.{_format}
+    defaults:
+        _format: html
+        _controller: schedule_assignor_controller_2016
+        _view_html:  schedule_assignor_view_html_2016
+        _view_xls:   schedule_assignor_view_xls_2016
+        _view_csv:   schedule_assignor_view_csv_2016
+        _view_txt:   schedule_assignor_view_txt_2016
+
 schedule_game_2016:
     path: /game.{_format}
     defaults:
